some flow control fixes moves TokenType to grammars namespace to resolve circular dependency
