+++ libraries/Pliant/Grammars/ILexerRule.cs
@@ -5,6 +5,7 @@ namespace Pliant.Grammars
     public interface ILexerRule : ISymbol
     {
         LexerRuleType LexerRuleType { get; }
+        // TODO: eliminate circular reference on Tokens namespace
         TokenType TokenType { get; }
         bool CanApply(char c);
     }
