+++ libraries/Pliant/Tokens/IToken.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Pliant.Grammars;
+using System.Collections.Generic;
 
 namespace Pliant.Tokens
 {
