+++ libraries/Pliant/Tokens/Token.cs
@@ -1,6 +1,6 @@
 ﻿using Pliant.Diagnostics;
+using Pliant.Grammars;
 using Pliant.Utilities;
-using System;
 using System.Collections.Generic;
 
 namespace Pliant.Tokens
