+++ tests/Pliant.Tests.Common/Forest/FakeTokenForestNode.cs
@@ -1,6 +1,7 @@
 ﻿using Pliant.Forest;
 using System;
 using Pliant.Tokens;
+using Pliant.Grammars;
 
 namespace Pliant.Tests.Common.Forest
 {
